Removing now works (calling EntityManager.refresh() instead EntityManager.persist() on the temporary drive), but obviously, the parameters at the remove link are wrong. I.e: not the waypoint that is clicked on gets removed, but (semingly) the one next in list.

+++ src/OpenRideServer/joride/src/java/de/avci/joride/jbeans/driverundertakesride/JWaypointEntity.java
@@ -261,7 +261,7 @@ public class JWaypointEntity extends WaypointEntity implements Serializable {
      *
      * @return this.getRouteIdx-0.5
      */
-    public double getPositionValueBefore() {
+    public int getPositionValueBefore() {
         return this.getRouteIdx();
     }
 
@@ -272,8 +272,8 @@ public class JWaypointEntity extends WaypointEntity implements Serializable {
      *
      * @return this.getRouteIdx+0.5
      */
-    public double getPositionValueAfter() {
-        return this.getRouteIdx() + 1;
+    public int getPositionValueAfter() {
+        return this.getRouteIdx()+1;
     }
